% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auptimizer.R
\name{get_config}
\alias{get_config}
\title{Get hyperparameter configuration}
\usage{
get_config(filename, env)
}
\arguments{
\item{filename}{optional JSON filename contains hyperparameters as dict, default to first argument in command line.}

\item{env}{optional environment contains hyperparameters, de fault to global env.}
}
\value{
A matrix of the infile
}
\description{
This function loads a file as a JSON. It assumes that the first argument of the running script is the configuraiton 
file.  It will parse and inject the hyperparameter variables into the global environment and return them.
}
