% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auptimizer.R
\name{get_config}
\alias{get_config}
\title{Get hyperparameter configuration}
\usage{
get_config(filename, env)
}
\arguments{
\item{filename}{optional JSON filename contains hyperparameters as dict, default to first argument in command line.}

\item{env}{optional environment contains hyperparameters, de fault to global env.}
}
\value{
A matrix of the infile
}
\description{
This function loads a configuration file in JSON format. It assumes that the first argument of the running script is
the configuraiton file.  It will parse and inject the hyperparameter variables into the global environment and return
them.
}
\details{
Assume an R script is used as `./test.R config.json`, the script should look like below:

\preformatted{
#!Rscript

# hyperparameter definitions
get_config() 
# use hyperparamter to train model, 
# after training, return optimization target (e.g. validation accuracy) as score
print_result(score)
}
}
