% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auptimizer.R
\name{print_result}
\alias{print_result}
\title{Report result to Auptimizer}
\usage{
print_result(score)
}
\arguments{
\item{score}{Score to be reported back to Auptimizer}
}
\description{
This function reports the results to Auptimizer.  A numerical score will be converted to string automatically.  If 
you want to report multiple scores, use ',' to seperate them and only the first value will be used by Auptimzier for
Optimization.
}
